backend/requirements.txt
openai>=1.0
fastapi>=0.104
uvicorn[standard]>=0.24
plotly>=5.17
python-dotenv>=1.0
pandas>=2.1
numpy>=1.24
pydantic>=2.0
